[SP2/Design] 헤더 및 에러 페이지 디자인 및 워딩 수정 (#279)
* feat: login 여부에 따른 라우팅 처리 달리하기

* chore: 불필요한 코드 제거

* design: error page 이전으로 돌아가기 버튼만 두기

* chore: header nav 내용 수정

* feat: my page footer 추가 및 about page 로고 클릭 비활성화

* refactor: my page 컴포넌트 분리

* design: my page footer border 제거

* fix: 서비스 소개 url 변경

* fix: 회원 탈퇴 및 로그아웃 시 라우팅 되는 페이지 url 변경
