+++ src/My/constants/FOOTER_LIST.ts
@@ -1,0 +1,20 @@
+export const FOOTER_LIST = [
+  {
+    label: '서비스 소개',
+    url: '/about',
+    newTab: false,
+    isBold: false,
+  },
+  {
+    label: '서비스 이용약관',
+    url: 'https://moonshot-.notion.site/moonshot-1-0-5068b03a06a245b2b3b89d1b406345fe?pvs=4',
+    newTab: true,
+    isBold: false,
+  },
+  {
+    label: '개인정보 처리방침',
+    url: 'https://moonshot-.notion.site/moonshot-1-0-0cdb3499765349e9aba64474fbba933e',
+    newTab: true,
+    isBold: true,
+  },
+];
