+++ src/Onboarding/constants/NAV_ITEMS.ts
@@ -1,10 +1,6 @@
 export const NAV_ITEMS = [
   {
-    text: '홈',
-    path: '/',
-  },
-  {
-    text: 'Team moonshot',
-    path: '/team-moonshot',
+    text: '서비스 소개',
+    path: '/about',
   },
 ];
